refactor: centralize #current-cloud-backend upload logic
