+++ packages/amplify-cli-core/src/errors/amplify-exception.ts
@@ -105,8 +105,8 @@ export type AmplifyErrorType =
   | 'ConfigurationError'
   | 'DeploymentError'
   | 'DeploymentInProgressError'
-  | 'DirectoryError'
   | 'DirectoryAlreadyExistsError'
+  | 'DirectoryError'
   | 'DuplicateLogicalIdError'
   | 'EnvironmentConfigurationError'
   | 'EnvironmentNameError'
@@ -124,10 +124,11 @@ export type AmplifyErrorType =
   | 'MigrationError'
   | 'MissingAmplifyMetaFileError'
   | 'MissingOverridesInstallationRequirementsError'
+  | 'MockProcessError'
   | 'ModelgenError'
   | 'NestedProjectInitError'
-  | 'NoUpdateBackendError'
   | 'NotImplementedError'
+  | 'NoUpdateBackendError'
   | 'OpenSslCertificateError'
   | 'ParameterNotFoundError'
   | 'PermissionsError'
@@ -145,36 +146,38 @@ export type AmplifyErrorType =
   | 'ResourceAlreadyExistsError'
   | 'ResourceInUseError'
   | 'ResourceNotReadyError'
+  | 'SearchableMockProcessError'
+  | 'SearchableMockUnavailablePortError'
+  | 'SearchableMockUnsupportedPlatformError'
   | 'StackNotFoundError'
   | 'StackStateError'
-  | 'UserInputError'
-  | 'MockProcessError'
-  | 'SearchableMockUnsupportedPlatformError'
-  | 'SearchableMockUnavailablePortError'
-  | 'SearchableMockProcessError';
+  | 'UserInputError';
 
 /**
  * Amplify fault types
  */
 export type AmplifyFaultType =
-  | 'AnalyticsCategoryFault'
   | 'AmplifyBackupFault'
+  | 'AnalyticsCategoryFault'
+  | 'AuthCategoryFault'
+  | 'BackendDeleteFault'
   | 'BackendPullFault'
   | 'ConfigurationFault'
-  | 'BackendDeleteFault'
   | 'ConfigurationFault'
   | 'DeploymentFault'
+  | 'DeploymentStateUploadFault'
+  | 'MockProcessFault'
   | 'NotificationsChannelAPNSFault'
   | 'NotificationsChannelEmailFault'
   | 'NotificationsChannelFCMFault'
-  | 'NotificationsChannelSmsFault'
   | 'NotificationsChannelInAppMessagingFault'
+  | 'NotificationsChannelSmsFault'
   | 'NotImplementedFault'
+  | 'PluginNotLoadedFault'
   | 'ProjectDeleteFault'
   | 'ProjectInitFault'
-  | 'PluginNotLoadedFault'
-  | 'PushResourcesFault'
   | 'PullBackendFault'
+  | 'PushResourcesFault'
   | 'ResourceExportFault'
   | 'ResourceNotFoundFault'
   | 'ResourceNotReadyFault'
@@ -187,6 +190,4 @@ export type AmplifyFaultType =
   | 'UnexpectedS3Fault'
   | 'UnknownFault'
   | 'UnknownNodeJSFault'
-  | 'MockProcessFault'
-  | 'AuthCategoryFault'
   | 'ZipExtractFault';
